Move get_max_id
